Reduce number of separate modal components
