--- litmus-portal/frontend/src/components/Modal/PasswordModal/styles.ts
@@ -1,42 +1,0 @@
-import { makeStyles, Theme } from '@material-ui/core/styles';
-
-const useStyles = makeStyles((theme: Theme) => ({
-  body: {
-    display: 'flex',
-    flexDirection: 'column',
-    justifyContent: 'center',
-    alignItems: 'center',
-    marginTop: theme.spacing(15),
-    paddingBottom: '7rem',
-  },
-  text: {
-    width: '23.5rem',
-    height: '5.875rem',
-    marginTop: theme.spacing(3.75),
-    marginBottom: theme.spacing(3.75),
-  },
-  typo: {
-    fontSize: '2.25rem',
-  },
-  text1: {
-    width: '27.5rem',
-    height: '1.6875rem',
-    marginBottom: theme.spacing(3.75),
-  },
-  typo1: {
-    fontSize: '1rem',
-  },
-  button: {
-    textTransform: 'none',
-    width: '9.25rem',
-    height: '2.75rem',
-    borderRadius: '0.1875rem',
-    marginTop: theme.spacing(3.75),
-    backgroundColor: theme.palette.secondary.dark,
-    color: theme.palette.common.white,
-    '&:hover': {
-      backgroundColor: theme.palette.secondary.dark,
-    },
-  },
-}));
-export default useStyles;
