+++ litmus-portal/frontend/src/components/Sections/Settings/UserManagementTab/CreateUser/NewUserModal/index.tsx
@@ -1,11 +1,11 @@
 import { Typography } from '@material-ui/core';
 import React from 'react';
 import { useSelector } from 'react-redux';
-import config from '../../../config';
-import { RootState } from '../../../redux/reducers';
-import ButtonFilled from '../../Button/ButtonFilled';
+import config from '../../../../../../config';
+import { RootState } from '../../../../../../redux/reducers';
+import ButtonFilled from '../../../../../Button/ButtonFilled';
 import useStyles from './styles';
-import Unimodal from '../../../containers/layouts/Unimodal';
+import Unimodal from '../../../../../../containers/layouts/Unimodal';
 
 // Props for NewUserModal component
 interface NewUserModalProps {
