+++ litmus-portal/frontend/src/components/Sections/Settings/UserManagementTab/CreateUser/NewUserModal/styles.ts
