+++ litmus-portal/frontend/src/components/WelcomeModal/styles.ts
